fix: move util package into ui package
