+++ app/src/main/java/com/neurosky/zonetrainer/ui/util/RecorderUtil.kt
@@ -1,4 +1,4 @@
-package com.neurosky.zonetrainer.util
+package com.neurosky.zonetrainer.ui.util
 
 import android.content.ContentValues
 import android.content.Context
