+++ app/src/main/java/com/neurosky/zonetrainer/ui/util/TgStreamHandlerImpl.kt
@@ -1,4 +1,4 @@
-package com.neurosky.zonetrainer.util
+package com.neurosky.zonetrainer.ui.util
 
 import android.util.Log
 import com.neurosky.connection.ConnectionStates
